fix(9a): CodeRabbitレビュー17件に対応する
PR #47への17件のCodeRabbit指摘を精査し、実際のバグ・堅牢性の欠落には
修正とテストを追加、設計意図と衝突する指摘は理由を付けて現状維持とした。

- findOrCreateCategory/findOrCreateTagをINSERT...ON CONFLICT DO UPDATE
  RETURNINGで冪等化し、SELECT-then-INSERTの競合を解消（categories.nameへ
  V8マイグレーションでUNIQUE制約を追加）
- JdbcPromptMetadataRepository.upsertをTransactionTemplateで1トランザクション化
  （DELETE後の失敗でタグが全消失する不具合を解消）
- JdbcPromptSearchRepository.queryItemsのfindTags呼び出しをN+1から
  バッチ取得（IN句1回）に変更し、ページサイズに比例したコネクション圧迫を解消
- JdbcAuditRepository.searchをLEFT JOIN化し、promptKeyがnullで記録された
  行（NIL_PROMPT_ID）もsearchで見つかるようにする。ORDER BYにaudit_idの
  タイブレーカーを追加し同一occurred_atでのページング順序を安定化
- PromptSearchCriteria/AuditQueryにpage/sizeのバリデーションを追加し、
  JDBCのOFFSET計算をLong演算に変更（オーバーフロー対策）
- InMemoryAuditRepositoryのページング計算もLong演算に変更
- PromptMetadataRepository.upsertのKDocを実装（Prompt Aggregate事前存在が
  前提、新規作成はしない）に合わせて修正。DependencyKind/find/searchへKDoc追加
- 実装ガイドのmarkdown lint警告を修正

AuditLogEntry.payloadをSensitiveValue化する指摘は見送った。KDocが既に
「呼出側がSecretマスク済であることを保証する」契約を明記しており、
payload自体は監査ログの表示内容そのもの（`GET /audit-logs`が返す値）
であるため、SensitiveValueでtoString()を"***"に固定すると監査ログ検索
機能自体が意味を失う。

dependencies.to_kind/execution_logs.statusのcasing不一致指摘は、前者は
PR #47内の既存修正（UPPER()比較）で解消済み、後者は本番書き込み経路が
まだ存在しないため各Repositoryのコードコメントで規約を明文化するに留めた。

+++ modules/prompt-engine-infrastructure/src/main/kotlin/promptengine/infrastructure/audit/InMemoryAuditRepository.kt
@@ -46,8 +46,11 @@ class InMemoryAuditRepository(activeProfiles: Set<String>) : AuditRepository {
                     .filter { query.to == null || !it.occurredAt.isAfter(query.to) }
                     .sortedByDescending { it.occurredAt }
             }
-        val fromIndex = (query.page * query.size).coerceAtMost(filtered.size)
-        val toIndex = (fromIndex + query.size).coerceAtMost(filtered.size)
+        // page/sizeはAuditQuery.initでpage>=0・size<=MAX_SIZEを保証されるが、pageの上限は
+        // 無いため、page * sizeをIntのまま計算するとオーバーフローしうる（CodeRabbitレビュー指摘）。
+        // Long演算で計算してからfiltered.sizeでクランプする。
+        val fromIndex = (query.page.toLong() * query.size).coerceAtMost(filtered.size.toLong()).toInt()
+        val toIndex = (fromIndex.toLong() + query.size).coerceAtMost(filtered.size.toLong()).toInt()
         return Page(
             items = filtered.subList(fromIndex, toIndex),
             page = query.page,
